feat: separate cmds into files
